# Conflicts:
#	assembly/pom.xml
#	atopcalcite/pom.xml
#	atopcalcite/src/main/java/org/apache/calcite/prepare/OnlyPrepareEarlyAbortException.java
#	core-common/pom.xml
#	core-common/src/main/java/org/apache/kylin/common/BackwardCompatibilityConfig.java
#	core-common/src/main/java/org/apache/kylin/common/KylinConfigBase.java
#	core-common/src/main/java/org/apache/kylin/common/KylinVersion.java
#	core-common/src/main/java/org/apache/kylin/common/persistence/BrokenEntity.java
#	core-common/src/main/java/org/apache/kylin/common/persistence/JDBCConnectionManager.java
#	core-common/src/main/java/org/apache/kylin/common/persistence/JDBCResourceDAO.java
#	core-common/src/main/java/org/apache/kylin/common/persistence/JDBCResourceStore.java
#	core-common/src/main/java/org/apache/kylin/common/persistence/JDBCSqlQueryFormat.java
#	core-common/src/main/java/org/apache/kylin/common/persistence/JDBCSqlQueryFormatProvider.java
#	core-common/src/main/java/org/apache/kylin/common/util/Bytes.java
#	core-common/src/main/java/org/apache/kylin/common/util/HiveCmdBuilder.java
#	core-common/src/main/java/org/apache/kylin/common/util/StringUtil.java
#	core-common/src/test/java/org/apache/kylin/common/util/InstallJarIntoMavenTest.java
#	core-cube/pom.xml
#	core-dictionary/pom.xml
#	core-job/pom.xml
#	core-job/src/main/java/org/apache/kylin/job/JoinedFlatTable.java
#	core-job/src/main/java/org/apache/kylin/job/impl/threadpool/DefaultScheduler.java
#	core-job/src/test/java/org/apache/kylin/job/PersistExceptionExecutable.java
#	core-metadata/pom.xml
#	core-metadata/src/main/java/org/apache/kylin/measure/MeasureTypeFactory.java
#	core-metadata/src/main/java/org/apache/kylin/measure/topn/TopNAggregator.java
#	core-metadata/src/main/java/org/apache/kylin/metadata/model/PartitionDesc.java
#	core-metadata/src/main/java/org/apache/kylin/metadata/model/TableDesc.java
#	core-metadata/src/test/java/org/apache/kylin/measure/AggregatorMemEstimateTest.java
#	core-metrics/pom.xml
#	core-storage/pom.xml
#	datasource-sdk/src/main/java/org/apache/kylin/sdk/datasource/framework/def/TypeMapping.java
#	datasource-sdk/src/test/java/org/apache/kylin/sdk/datasource/framework/SourceConnectorFactoryTest.java
#	engine-mr/pom.xml
#	engine-spark/pom.xml
#	engine-spark/src/main/java/org/apache/kylin/engine/spark/SparkCubingByLayer.java
#	engine-spark/src/main/java/org/apache/kylin/engine/spark/SparkCubingMerge.java
#	engine-spark/src/main/java/org/apache/kylin/engine/spark/SparkFactDistinct.java
#	jdbc/pom.xml
#	kylin-it/pom.xml
#	kylin-it/src/test/java/org/apache/kylin/provision/BuildCubeWithEngine.java
#	kylin-it/src/test/java/org/apache/kylin/storage/jdbc/ITJDBCResourceStoreTest.java
#	metrics-reporter-hive/pom.xml
#	metrics-reporter-kafka/pom.xml
#	pom.xml
#	query/pom.xml
#	server-base/pom.xml
#	server-base/src/main/java/org/apache/kylin/rest/controller/CubeController.java
#	server-base/src/main/java/org/apache/kylin/rest/controller/ModelController.java
#	server-base/src/main/java/org/apache/kylin/rest/controller/ProjectController.java
#	server-base/src/main/java/org/apache/kylin/rest/service/ProjectService.java
#	server-base/src/main/java/org/apache/kylin/rest/service/QueryService.java
#	server/pom.xml
#	source-hive/pom.xml
#	source-hive/src/main/java/org/apache/kylin/source/hive/GarbageCollectionStep.java
#	source-hive/src/main/java/org/apache/kylin/source/hive/HiveInputBase.java
#	source-hive/src/main/java/org/apache/kylin/source/hive/HiveMetadataExplorer.java
#	source-hive/src/main/java/org/apache/kylin/source/hive/cardinality/HiveColumnCardinalityJob.java
#	source-jdbc/src/main/java/org/apache/kylin/source/jdbc/JdbcHiveMRInput.java
#	source-kafka/pom.xml
#	storage-hbase/pom.xml
#	storage-hbase/src/main/java/org/apache/kylin/storage/hbase/steps/CreateHTableJob.java
#	storage-hbase/src/main/java/org/apache/kylin/storage/hbase/steps/SparkCubeHFile.java
#	storage-hbase/src/main/java/org/apache/kylin/storage/hbase/util/GridTableHBaseBenchmark.java
#	storage-hbase/src/main/java/org/apache/kylin/storage/hbase/util/HbaseStreamingInput.java
#	storage-hbase/src/test/java/org/apache/kylin/storage/hbase/common/HiveJDBCClientTest.java
#	tomcat-ext/pom.xml
#	tool-assembly/pom.xml
#	tool/pom.xml
#	tool/src/main/java/org/apache/kylin/tool/AbstractInfoExtractor.java
#	webapp/app/js/controllers/cubeMeasures.js

+++ engine-mr/src/main/java/org/apache/kylin/engine/mr/common/AbstractHadoopJob.java
@@ -324,7 +324,7 @@ public abstract class AbstractHadoopJob extends Configured implements Tool {
             StringBuilder jarList = new StringBuilder();
             StringBuilder fileList = new StringBuilder();
 
-            for (String fileName : kylinDependency.split(",")) {
+            for (String fileName : StringUtil.splitAndTrim(kylinDependency, ",")) {
                 Path p = new Path(fileName);
                 if (p.isAbsolute() == false) {
                     logger.warn("The directory of kylin dependency '" + fileName + "' is not absolute, skip");
@@ -625,24 +625,29 @@ public abstract class AbstractHadoopJob extends Configured implements Tool {
     }
 
     protected void cleanupTempConfFile(Configuration conf) {
-        String tempMetaFileString = conf.get("tmpfiles");
-        logger.trace("tempMetaFileString is : " + tempMetaFileString);
-        if (tempMetaFileString != null) {
-            if (tempMetaFileString.startsWith("file://")) {
-                tempMetaFileString = tempMetaFileString.substring("file://".length());
-                File tempMetaFile = new File(tempMetaFileString);
-                if (tempMetaFile.exists()) {
-                    try {
-                        FileUtils.forceDelete(tempMetaFile.getParentFile());
+        String[] tempfiles = StringUtils.split(conf.get("tmpfiles"), ",");
+        if (tempfiles == null) {
+            return;
+        }
+        for (String tempMetaFileString : tempfiles) {
+            logger.trace("tempMetaFileString is : " + tempMetaFileString);
+            if (tempMetaFileString != null) {
+                if (tempMetaFileString.startsWith("file://")) {
+                    tempMetaFileString = tempMetaFileString.substring("file://".length());
+                    File tempMetaFile = new File(tempMetaFileString);
+                    if (tempMetaFile.exists()) {
+                        try {
+                            FileUtils.forceDelete(tempMetaFile.getParentFile());
 
-                    } catch (IOException e) {
-                        logger.warn("error when deleting " + tempMetaFile, e);
+                        } catch (IOException e) {
+                            logger.warn("error when deleting " + tempMetaFile, e);
+                        }
+                    } else {
+                        logger.info("" + tempMetaFileString + " does not exist");
                     }
                 } else {
-                    logger.info("" + tempMetaFileString + " does not exist");
+                    logger.info("tempMetaFileString is not starting with file:// :" + tempMetaFileString);
                 }
-            } else {
-                logger.info("tempMetaFileString is not starting with file:// :" + tempMetaFileString);
             }
         }
     }
